Refactor project structure to yarn workspaces
