+++ api/nodemon.json
@@ -1,0 +1,5 @@
+{
+  "watch": ["./src/"],
+  "ext": ".ts,.js",
+  "exec": "ts-node -r dotenv/config ./src/server.ts dotenv_config_path=../.env "
+}
