+++ api/package.json
@@ -1,0 +1,64 @@
+{
+  "name": "api",
+  "version": "1.0.0",
+  "description": "The api part of the monorepo.",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Tebra/node-vue3-typescript.git"
+  },
+  "engines": {
+    "node": "12.x"
+  },
+  "author": "Rijad Zuzo",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/Tebra/node-vue3-typescript/issues"
+  },
+  "homepage": "https://github.com/Tebra/node-vue3-typescript#readme",
+  "scripts": {
+    "build": "rimraf ../build && tsc",
+    "serve": "nodemon",
+    "test": "mocha -r ts-node/register ./src/tests/**/*.ts",
+    "format": "prettier --write ./src"
+  },
+  "dependencies": {
+    "compression": "^1.7.4",
+    "cors": "^2.8.5",
+    "dotenv": "^10.0.0",
+    "express": "^4.17.1",
+    "express-winston": "^4.2.0",
+    "helmet": "^4.6.0",
+    "pg": "^8.7.1",
+    "pg-hstore": "^2.3.4",
+    "reflect-metadata": "^0.1.13",
+    "sequelize": "6.6.2",
+    "sequelize-typescript": "^2.1.0",
+    "winston": "^3.3.3"
+  },
+  "devDependencies": {
+    "@types/chai": "^4.2.11",
+    "@types/mocha": "^5.2.4",
+    "@types/compression": "^1.7.1",
+    "@types/cors": "^2.8.12",
+    "@types/express": "^4.17.13",
+    "@types/faker": "^5.5.8",
+    "@types/helmet": "^4.0.0",
+    "@types/node": "^16.7.4",
+    "@types/sequelize": "^4.28.10",
+    "@typescript-eslint/eslint-plugin": "^4.18.0",
+    "@typescript-eslint/parser": "^4.18.0",
+    "@vicons/tabler": "^0.11.0",
+    "chai": "^4.1.2",
+    "eslint": "^6.7.2",
+    "eslint-plugin-prettier": "^3.3.1",
+    "eslint-plugin-vue": "^7.0.0",
+    "faker": "^5.5.3",
+    "mocha": "^9.1.1",
+    "nodemon": "^2.0.12",
+    "prettier": "2.3.2",
+    "rimraf": "^3.0.2",
+    "sass-loader": "^8.0.2",
+    "ts-node": "^10.2.1",
+    "typescript": "~4.1.5"
+  }
+}
