--- nodemon.json
@@ -1,10 +1,0 @@
-{
-  "watch": [
-    "./api/"
-  ],
-  "ext": ".ts,.js",
-  "ignore": [
-    "./client"
-  ],
-  "exec": "ts-node -r dotenv/config ./api/server.ts"
-}
