+++ package.json
@@ -9,6 +9,12 @@
   "engines": {
     "node": "12.x"
   },
+  "workspaces": [
+    "api",
+    "shared",
+    "ui"
+  ],
+  "private": true,
   "keywords": [
     "Node",
     "Vue",
@@ -22,83 +28,23 @@
   },
   "homepage": "https://github.com/Tebra/node-vue3-typescript#readme",
   "scripts": {
-    "serve:api": "nodemon",
-    "serve:client": "vue-cli-service serve --port 8080 --open",
-    "serve": "npm-run-all --parallel serve:api serve:client",
-    "test:api": "mocha -r ts-node/register ./api/tests/**/*.ts",
-    "test:client": "vue-cli-service test:unit ./client/tests/**/*.ts",
-    "build:api": "rimraf ./build && tsc",
-    "build:client": "vue-cli-service build --dest ./build/public",
-    "build": "npm-run-all build:api build:client",
-    "format:api": "prettier --write ./api",
-    "format:client": "prettier --write ./client",
-    "format": "npm-run-all format:api format:client",
+    "serve:api": "yarn workspace api serve",
+    "serve:ui": "yarn workspace ui serve",
+    "serve": "npm-run-all --parallel serve:api serve:ui",
+    "test:api": "yarn workspace api test",
+    "test:ui": "yarn workspace ui test",
+    "test": "npm-run-all test:api test:ui",
+    "build:api": "yarn workspace api build",
+    "build:ui": "yarn workspace ui build",
+    "build": "npm-run-all build:api build:ui",
+    "format:api": "yarn workspace api format",
+    "format:ui": "yarn workspace ui build",
+    "format": "npm-run-all format:api format:ui",
     "start": "node -r dotenv/config ./build/server.js",
-    "lint": "vue-cli-service lint",
+    "lint": "yarn workspace ui lint",
     "trace": "tsc --traceResolution"
   },
-  "dependencies": {
-    "axios": "^0.21.4",
-    "compression": "^1.7.4",
-    "core-js": "^3.6.5",
-    "cors": "^2.8.5",
-    "dotenv": "^10.0.0",
-    "express": "^4.17.1",
-    "express-winston": "^4.2.0",
-    "helmet": "^4.6.0",
-    "pg": "^8.7.1",
-    "pg-hstore": "^2.3.4",
-    "pinia": "^2.0.0-rc.9",
-    "reflect-metadata": "^0.1.13",
-    "register-service-worker": "^1.7.1",
-    "sequelize": "6.6.2",
-    "sequelize-typescript": "^2.1.0",
-    "vue": "^3.0.0",
-    "vue-class-component": "^8.0.0-0",
-    "vue-router": "^4.0.0-0",
-    "winston": "^3.3.3"
-  },
   "devDependencies": {
-    "@types/chai": "^4.2.11",
-    "@types/compression": "^1.7.1",
-    "@types/cors": "^2.8.12",
-    "@types/express": "^4.17.13",
-    "@types/faker": "^5.5.8",
-    "@types/helmet": "^4.0.0",
-    "@types/mocha": "^5.2.4",
-    "@types/node": "^16.7.4",
-    "@types/sequelize": "^4.28.10",
-    "@typescript-eslint/eslint-plugin": "^4.18.0",
-    "@typescript-eslint/parser": "^4.18.0",
-    "@vicons/tabler": "^0.11.0",
-    "@vue/cli-plugin-babel": "~4.5.0",
-    "@vue/cli-plugin-eslint": "~4.5.0",
-    "@vue/cli-plugin-pwa": "~4.5.0",
-    "@vue/cli-plugin-router": "~4.5.0",
-    "@vue/cli-plugin-typescript": "~4.5.0",
-    "@vue/cli-plugin-unit-mocha": "~4.5.0",
-    "@vue/cli-plugin-vuex": "~4.5.0",
-    "@vue/cli-service": "~4.5.0",
-    "@vue/compiler-sfc": "^3.0.0",
-    "@vue/eslint-config-prettier": "^6.0.0",
-    "@vue/eslint-config-typescript": "^7.0.0",
-    "@vue/test-utils": "^2.0.0-0",
-    "chai": "^4.1.2",
-    "eslint": "^6.7.2",
-    "eslint-plugin-prettier": "^3.3.1",
-    "eslint-plugin-vue": "^7.0.0",
-    "faker": "^5.5.3",
-    "mocha": "^9.1.1",
-    "naive-ui": "^2.18.2",
-    "node-sass": "^4.12.0",
-    "nodemon": "^2.0.12",
-    "npm-run-all": "^4.1.5",
-    "prettier": "2.3.2",
-    "rimraf": "^3.0.2",
-    "sass-loader": "^8.0.2",
-    "ts-node": "^10.2.1",
-    "tsconfig-paths-webpack-plugin": "^3.5.1",
-    "typescript": "~4.1.5",
-    "vfonts": "^0.1.0"
+   "npm-run-all": "^4.1.5"
   }
 }
