+++ tsconfig.json
@@ -1,20 +1,19 @@
 {
   "compilerOptions": {
-    "target": "es6",
+    /* Basic Options */
+    "target": "es5",
     "module": "commonjs",
-    "lib": ["es6"],
-    "allowJs": true,
-    "outDir": "./build",
-    "rootDir": "./api",
-    "baseUrl": ".",
+    "lib": ["ESNext"],
+
+    /* Strict Type-Checking Options */
     "strict": true,
-    "noImplicitAny": true,
+
+    /* Module Resolution Options */
+    "moduleResolution": "node",
     "esModuleInterop": true,
-    "resolveJsonModule": true,
-    "experimentalDecorators": true,
-    "emitDecoratorMetadata": true,
-    "typeRoots": ["./node_modules/@types"]
-  },
-  "include": ["./api/**/*"],
-  "exclude": ["node_modules", "client", "./api/tests/**/*"]
+
+    /* Advanced Options */
+    "forceConsistentCasingInFileNames": true,
+    "skipLibCheck": true,
+  }
 }
