+++ ui/package.json
@@ -1,0 +1,59 @@
+{
+  "name": "ui",
+  "version": "1.0.0",
+  "description": "The ui part of the monorepo",
+  "author": "Rijad Zuzo",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/Tebra/node-vue3-typescript/issues"
+  },
+  "homepage": "https://github.com/Tebra/node-vue3-typescript#readme",
+  "scripts": {
+    "build": "vue-cli-service build --dest ../build/public",
+    "serve": "vue-cli-service serve --port 8080 --open",
+    "test": "vue-cli-service test:unit ./src/tests/**/*.ts",
+    "format": "prettier --write ./src",
+    "lint": "vue-cli-service lint"
+  },
+  "dependencies": {
+    "axios": "^0.21.4",
+    "core-js": "^3.6.5",
+    "pinia": "^2.0.0-rc.9",
+    "reflect-metadata": "^0.1.13",
+    "register-service-worker": "^1.7.1",
+    "vue": "^3.0.0",
+    "vue-class-component": "^8.0.0-0",
+    "vue-router": "^4.0.0-0"
+  },
+  "devDependencies": {
+    "@types/chai": "^4.2.11",
+    "@types/mocha": "^5.2.4",
+    "@typescript-eslint/eslint-plugin": "^4.18.0",
+    "@typescript-eslint/parser": "^4.18.0",
+    "@vicons/tabler": "^0.11.0",
+    "@vue/cli-plugin-babel": "~4.5.0",
+    "@vue/cli-plugin-eslint": "~4.5.0",
+    "@vue/cli-plugin-pwa": "~4.5.0",
+    "@vue/cli-plugin-router": "~4.5.0",
+    "@vue/cli-plugin-typescript": "~4.5.0",
+    "@vue/cli-plugin-unit-mocha": "~4.5.0",
+    "@vue/cli-plugin-vuex": "~4.5.0",
+    "@vue/cli-service": "~4.5.0",
+    "@vue/compiler-sfc": "^3.0.0",
+    "@vue/eslint-config-prettier": "^6.0.0",
+    "@vue/eslint-config-typescript": "^7.0.0",
+    "@vue/test-utils": "^2.0.0-0",
+    "chai": "^4.1.2",
+    "eslint": "^6.7.2",
+    "eslint-plugin-prettier": "^3.3.1",
+    "eslint-plugin-vue": "^7.0.0",
+    "mocha": "^9.1.1",
+    "naive-ui": "^2.18.2",
+    "node-sass": "^4.12.0",
+    "prettier": "2.3.2",
+    "sass-loader": "^8.0.2",
+    "tsconfig-paths-webpack-plugin": "^3.5.1",
+    "typescript": "~4.1.5",
+    "vfonts": "^0.1.0"
+  }
+}
