+++ ui/src/registerServiceWorker.ts
